Update OneHitWin.m
Considering the case where dealer has an Ace, when he/she draws a card that will burst, the Ace will become value of 1 to avoid bursting. But the maximum condition is that the dealer has an Ace and a 5 with value of 16 and hit one time, if he/she draws a 10/J/Q/K, the card value will become 16 again and he/she needs to draw one more time. This condition has not been considered yet.